Move org.eclipse.leshan in org.eclipse.leshan.core for src/test/java too
