+++ leshan-core/src/test/java/org/eclipse/leshan/core/json/JsonDeserializerTest.java
@@ -14,7 +14,7 @@
  *     Sierra Wireless - initial API and implementation
  *     Gemalto M2M GmbH
  *******************************************************************************/
-package org.eclipse.leshan.json;
+package org.eclipse.leshan.core.json;
 
 import org.eclipse.leshan.core.json.JsonRootObject;
 import org.eclipse.leshan.core.json.LwM2mJson;
