+++ leshan-core/src/test/java/org/eclipse/leshan/core/json/JsonSerializerTest.java
@@ -14,7 +14,7 @@
  *     Sierra Wireless - initial API and implementation
  *     Gemalto M2M GmbH
  *******************************************************************************/
-package org.eclipse.leshan.json;
+package org.eclipse.leshan.core.json;
 
 import java.util.ArrayList;
